pyproject.toml
[tool.poetry]
name = "geheimvz"
version = "1.0.0"
description = "Slow Social Network"
authors = ["Leonard Techel <git@barnslig.eu>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
django-environ = "^0.11.2"
django-vite = "^3.0.5"
django-crispy-forms = "^2.3"
crispy-tailwind = "^1.0.3"
pillow = "^11.0.0"
whitenoise = {extras = ["brotli"], version = "^6.8.2"}
django-storages = {extras = ["s3"], version = "^1.14.4"}
django-debug-toolbar = "^4.4.6"
django-cleanup = "^9.0.0"
django-friendship = "^1.9.6"
django-tables2 = "^2.7.0"
daphne = "^4.1.2"
django-ratelimit = "^4.1.0"
django-imagekit = "^5.0.0"
pillow-avif-plugin = "^1.4.6"
django-pictures = "^1.3.3"
filetype = "^1.2.0"
django-components = "^0.102"
django-csp = "^3.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
